@robodoo 2025-08-04
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/221724.png)](https://mergebot.odoo.com/odoo/odoo/pull/221724)